+++ public/data/product-detail.json
@@ -16,5 +16,5 @@
   "meat": "양고기",
   "eng_meat": "a",
   "food_image": "afdas",
-  "review": null
+  "review": "JMT입니다. 완전 강추에요. 이렇게 맛있는 베이에나는 태어나서 처음 먹어봐요!! 사실 베이에나를 처음 먹어봐요 :)"
 }
