+++ src/pages/ProductDetail/component/AddCart/AddCart.scss
@@ -1,0 +1,50 @@
+@import '/src/styles/variables';
+@import '/src/styles/color';
+
+.addcart {
+  .product-total-price {
+    display: flex;
+    align-items: end;
+    justify-content: end;
+    padding-top: 20px;
+    gap: 8px;
+
+    .total-price-text {
+      @include body-2;
+      height: 100%;
+    }
+
+    .total-price-number {
+      @include head-1;
+    }
+  }
+
+  .button-box {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding-top: 20px;
+
+    .recommand-button {
+      width: 60px;
+      height: 60px;
+      border-radius: 50%;
+      &:hover {
+        background-color: #f5f5f5;
+        cursor: pointer;
+      }
+    }
+    .cart-add-button {
+      width: 300px;
+      height: 66px;
+      border-radius: 30px;
+      background-color: $main-dark;
+      color: $main-white;
+      @include body-2;
+      &:hover {
+        opacity: 0.5;
+        cursor: pointer;
+      }
+    }
+  }
+}
