--- src/pages/ProductDetail/component/ProductImformation.scss
@@ -1,31 +1,0 @@
-@import '/src/styles/color';
-@import '/src/styles/variables';
-
-.product-information {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding-top: 60px;
-
-  .description-button-box {
-    .description-button {
-      width: 330px;
-      height: 50px;
-      border: 1px solid$main-grey;
-      border-radius: 0;
-      &:hover {
-        opacity: 0.5;
-        cursor: pointer;
-      }
-    }
-  }
-
-  .info-main-container {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    .description-img {
-      width: 1300px;
-    }
-  }
-}
