+++ src/pages/ProductDetail/component/ProductInformation/MainInfo/MainInfo.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import './MainInfo.scss';
+
+const MainInfo = () => {
+  return (
+    <div className="maininfo">
+      <img
+        className="maininfo-img"
+        src="/images/product/상품설명.png"
+        alt="상품설명"
+      />
+    </div>
+  );
+};
+
+export default MainInfo;
