+++ src/pages/ProductDetail/component/ProductInformation/MainInfo/MainInfo.scss
@@ -1,0 +1,10 @@
+@import '/src/styles/variables';
+@import '/src/styles/color';
+
+.maininfo {
+  width: 100%;
+
+  .maininfo-img {
+    width: 1020px;
+  }
+}
