+++ src/pages/ProductDetail/component/ProductInformation/ProductImformation.scss
@@ -1,0 +1,38 @@
+@import '/src/styles/color';
+@import '/src/styles/variables';
+
+.product-information {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 150px 0;
+
+  .description-button-box {
+    position: sticky;
+    top: 100px;
+    .description-button {
+      width: 340px;
+      height: 60px;
+      border: 1px solid lightgray;
+      background-color: #f7f7f7;
+      @include body-2;
+      &:hover {
+        opacity: 0.5;
+        cursor: pointer;
+      }
+    }
+  }
+
+  .info-main-container {
+    @include flexCenter;
+    flex-direction: column;
+
+    .scroll-box {
+      padding-top: 160px;
+
+      &:first-child {
+        padding: 0;
+      }
+    }
+  }
+}
