+++ src/pages/ProductDetail/component/ProductInformation/SubInfo/SubInfo.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import './SubInfo.scss';
+
+const SubInfo = () => {
+  return (
+    <div className="subinfo">
+      <img
+        src="/images/product/상세정보.png"
+        alt="상세정보"
+        className="subinfo-img"
+      />
+    </div>
+  );
+};
+
+export default SubInfo;
