+++ src/pages/ProductDetail/component/ProductInformation/SubInfo/SubInfo.scss
@@ -1,0 +1,10 @@
+@import '/src/styles/variables';
+@import '/src/styles/color';
+
+.subinfo {
+  width: 100%;
+
+  .subinfo-img {
+    width: 1020px;
+  }
+}
